Release v3.1.9 beta 3: Universal Triggers manager window, slide-out editor, status pills, and standard managed reward sync

+++ VrcTwitchOscBridge/Services/BridgeRuntimeConfiguration.cs
@@ -121,7 +121,7 @@ public sealed record TriggerRuleSnapshot(
     IReadOnlyList<string> AvatarRouletAvatarNames,
     int RangeMinimum,
     int RangeMaximum,
-    int DurationSeconds,
+    double DurationSeconds,
     int CooldownSeconds,
     bool UsesLinkedChannelPointReward,
     int? BotMessageCooldownSeconds,
